requirements.txt
langchain
faiss-cpu
openai
sentence-transformers
python-dotenv
fastapi
uvicorn
